Improves the generated diff
We are removing the lines numbers from the codeSnip we shared to the
model, this helps to decrease chance of the diff adding line numbers to
each line it generates.

We still see some diffs not in an expected patch format, it seems like
if we dont provide an example diff then our output may not be in the
format we want.